refactor: 게임 로직 모듈화
