+++ src/user/askMaxAttempts.js
@@ -10,7 +10,6 @@ async function askMaxAttempts() {
       return maxAttempts;
     } catch (error) {
       console.log(error.message);
-      console.log("다시 입력해주세요.");
     }
   }
 }
