+++ src/user/askUserGuess.js
@@ -5,7 +5,6 @@ async function askUserGuess() {
       return validateNumber(guessInput);
     } catch (error) {
       console.log(error.message);
-      console.log("다시 입력해주세요.");
     }
   }
 }
